Shift kafka logic from controller to service
